+++ doc/whats_new.rst
@@ -87,6 +87,9 @@ Changelog
    - Fixed bug in :class:`linear_model.stochastic_gradient` :
      ``l1_ratio`` was used as ``(1.0 - l1_ratio)`` .
 
+   - Fixed bug in :class:`multiclass.OneVsOneClassifier` with string
+     labels
+
 API changes summary
 -------------------